Initial Commit (READ DESCRIPTION!)
This is all of Dr. Bernstein's classes. PA5 & PA6 classes are my interpretations. They can easily be replaced if Bernstein's classes are up or if someone has a better implementation.

Anyways, in the git-ignore file, I had to remove .str, .geo, .gps files to be able to commit. These files are really large. Just download them and have them on your local machine.