Fixed Redundant Code & Map Matching
In GPGGASentence, it was double correcting if the value was S or W. Therefore, the negative it should have been was being cancelled.

For map matching, this was more thinking about the problem itself. I put a comment in DynamicCartographyPanel if anyone cares to read.

I also created the simulator package which is a way to test PA6 without GPS coords.